Ruby module refactor, binding layer cleanups, allow DisplayString override for KLists

+++ modules/python/k_python_list.cpp
@@ -77,7 +77,7 @@ namespace kroll
 		}
 	}
 
-	void KPythonList::SetAt(unsigned long index, SharedValue value)
+	void KPythonList::SetAt(unsigned int index, SharedValue value)
 	{
 		while (index >= this->Size())
 		{
